Avances, subiendo para ver si se borran errores de core
